Refactor to support loading from string.
Refactor the keyfile code to provide both a `load_key` which reads a
keyfile, and `from_base64`, which loads directly from a base64 string.
This is to support the 'repokey' mode where the key is embedded
directly in the repo.

FossilOrigin-Name: d87bef7ff16a21ffdca85bdfe1071ef50308764c775882e54e7861687d423020