Basic repo support.
This can open repos configured with either repokey or keyfile.  It
will use BORG_PASSPHRASE if it is set, otherwise, it assumes an empty
passphrase.  The main accepts the path to the repo, and prints out the
basics of the repo.

FossilOrigin-Name: 0669ccc58fd0d607779b6e0693066269712b4420ad23bd93e60e770c33fa6821