Minimum code needed to test a borg-created repo.
Several things need to change to make this work.  1. The BORG_BASE
needs to be settable, and 2. The BORG_PASSWORD need to be set.
Because ocaml doesn't write to the environment, these have special
setters.  Unfortunately, this makes the tests non thread-safe.

Only 'repokey' works.  For 'keyfile', the name of the generated key,
is incorrect.  For 'none', it still tries loading the keyfile, even
though there is no need to do so.  The other modes, such as
authenticated, and Blake2 are not yet supported, either.

FossilOrigin-Name: 45f07f70996bd95505c675f227b5719e0aa983844815b7cf2f8799f8cb421d27